A treatise on the myriad ways to obtain url paramaters
Getting asynchronous data should be done on the onInit. Constructor is
for depenedency injection and setting up initial props.

The active Route on the component allows us to see the current url.
We can use snapshot to obtain a non observable (snapshot) of our data.
Or we can use the observable pattern which will listen to all changes to
active route. In this use case where the view should always proceed with
data from an immediate action, Observables her is over kill. Observables
are used to centralize code and to respond to changes from other components
such as managing global modal states.