Hook up product edit view to edit changes.
Here the navbar has a button that will change the params. Expected behavior
is to change the view. We could change the inital product property of the
component via a button on the current component but the navbar only has
access to changes to params. This is a good time to use Observables
so we can have the function to initialize this.product refire on params
change.