Allow the filtered results to be shown after clicking away
The filter query and is Image shown on product List are state.
They are brough over by the activated link on the product name.
As such they are brough over as query params. Pressing and activating
the back button link on our product detail view should retain these params
over to the new path /products. We do this with queryParamsHandling

In the view we set the `queryParamsHandling="preserve"` as an attribute
in the js we do
```
this.router.navigate(['product'], {
  queryParamsHandling: true
})
```

In additional, when the query params are preserved, the product list must
listen to these changes and change the initial props. ngzone will see that
the starter proteins are scarred and generate renders based on that value.
filter subscripte, then populate our listfilter and showImages starterproteins

A good way to grok the get and set pattern they use for listfilter is that
it's a good way to remove the (). Think of it as a .length property on an
array.