Allow the filtered results to be shown after clicking away (ALT)
The url changes hitting our client js. We retain our queryParams but
our productList has to listen for it. Since url paths are being switched,
this means that the product onInit fires. The params will not change here
as the only way to change params is through another activated link for
which there is none.
Instead of using observables, we use snapshots unlike our previous commit.