Create a product resolver service
Reference Personal Notes
[Resolver Notes](:note:e4d8cbd4-e575-4f00-858c-f13a987ee815)
This resolver is based on intercepting the request for data via the activated
Link and the component. This service is placed in core because of
https://itnext.io/choosing-a-highly-scalable-folder-structure-in-angular-d987de65ec7
where it explains how to create a scalable directory for angular
* “Structure the app such that you can locate code quickly”
*  “Have a near-term view of implementation and a long-term vision.
* Start small but keep in mind where the app is heading down the road”.

The SharedModule is where any shared components,
* pipes/filters
* services should go.
The SharedModule can be imported in any other module when those items will be re-used.
The shared module shouldn’t have any dependency to the rest of the application
and should therefore not rely on any other module.

The CoreModule takes on the role of the root AppModule ,
but is not the module which gets bootstrapped by Angular at run-time.
The CoreModule should contain singleton services (which is usually the case),
universal components and other features where there’s only once instance
 per application. To prevent re-importing the core module elsewhere,
 you should also add a guard for it in the core module’ constructor.

 This resolver is a singleton service as will a majority of the services
 you choose.