In the previous commit, we created a product Resolver. But we changed the return value to the generic Product Resolved. { product, error}.
If we look at the product module, we see that the the routing has to
go through a resolve propety. Like graph ql, a resolver Class resolves all
the data into resolvedData.

We assume that the ResolverClass has done it's mapping job.
with bundles it into our product error object stream.

In our detail component, the data is already inside the snapshot data property.
Accessing the data array from the activated route service gives us a reference
to the data instance. All code that changes that resolveData array is shared
with anyone that is linked via this.activeroute.snashop.data...
All folks linked to the same activeRoute will be privy to that datachange.
Our resolver works with a stored activated Route snapshot. That's our source.