Run validation edge cases with product module
Reset the edit component so that the edit form is clear off validation
errors on initial start. // TODO why does the form validation carry over?

Introduce the validation paradigm where the parent can activate or deactivate
a button based on the child routes.
Our saveproduct is only allowed to save if isValid return true.
A child route manifest is stored in dataIsValid. isValid will run a validate
command that will check on each validation error each route is concerned with
for info, this is productname and product code.
for tags, it is the category length.

When we run a blank isValid, it checks every key to be true and only returns true
when it is. The disabled attribute on the template is bound to the isValid
function.